Basic implementation of checklist card component
